Programmatic support for token transformer (#1514)
This exposes the transformTokens method to be used programmatically
instead of having to only use the CLI approach. It also fixes a minor
bug found with windows machines where the path seperator was treated
differently causing tests to fail

Fixes https://github.com/tokens-studio/figma-plugin/issues/1259
