+++ token-transformer/README.md
@@ -2,7 +2,9 @@
 
 Converts tokens from Tokens Studio for Figma to something Style Dictionary can read, removing any math operations or aliases, only resulting in raw values.
 
-## How to use
+## CLI usage
+
+### How to use
 
 Install (either globally or local)
 `npm install token-transformer -g`
@@ -25,7 +27,7 @@ You can also set a directory as an input instead of providing just one file.
 
 `node token-transformer src output.json core/colors,themes/dark core/colors`
 
-## Parameters
+### Parameters
 
 Input: Filename of input
 
@@ -44,3 +46,50 @@ Excludes: Sets that should not be part of the export (e.g. a global color scale)
 --throwErrorWhenNotResolved: true|false to enable/disable throwing errors when a reference fails to resolve (default: false)
 
 --resolveReferences: true|false|'math' to enable/disable resolving references, removing any aliases or math expressions (default: true)
+
+## Programmatic usage
+
+This library can also be used programmatically to resolve tokens without giving it access to the underlying file system.
+
+```js
+const { transformTokens } = require('token-transformer');
+
+const rawTokens = {
+    setA:{
+        "sizing": {
+            "base": {
+            "value": "4",
+            "description": "Alias value",
+            "type": "sizing"
+            },
+            "large": {
+            "value": "$sizing.base * 2",
+            "description": "Math value",
+            "type": "sizing"
+            }
+        }
+    }
+};
+
+const setsToUse = ['setA'];
+const excludes = [];
+
+const transformerOptions = {
+  expandTypography: true,
+  expandShadow: true,
+  expandComposition: true,
+  preserveRawValue: false,
+  throwErrorWhenNotResolved:  true,
+  resolveReferences:true
+}
+
+const resolved = transformTokens(rawTokens,setsToUse, excludes,transformerOptions);
+
+/*{
+  sizing: {
+    base: { value: 4, description: 'Alias value', type: 'sizing' },
+    large: { value: 8, description: 'Math value', type: 'sizing' } 
+  }
+}*/
+
+```
