+++ token-transformer/package.json
@@ -9,25 +9,26 @@
   },
   "main": "index.js",
   "scripts": {
-    "build-typography-test": "node index.js tokens.json temp/typography.json base/options,base/sizing,core/typography base/options,base/sizing",
-    "build-composition-test": "node index.js tokens.json temp/composition.json base/options,base/sizing,core/composition base/options,base/sizing --expandComposition=true",
-    "build-light-test": "node index.js tokens.json temp/dark.json base/options,base/sizing,themes/dark,core/shadows base/options,base/sizing",
-    "build-dark-test": "node index.js tokens.json temp/light.json base/options,base/sizing,themes/light,core/shadows base/options,base/sizing",
-    "build-single-file-test-theme-config": "node index.js tokens.json --theme --themeOutputPath temp",
-    "build-single-file-test-order-config": "node index.js order.json --theme --themeOutputPath temp",
-    "build-folder-test": "node index.js tokens temp/folder-light.json base/options,base/sizing,themes/light,core/shadows base/options,base/sizing",
-    "build-folder-test-theme-config": "node index.js tokens --theme --themeOutputPath temp",
-    "build-resolve-false-test": "node index.js tokens.json temp/resolve-false.json base/options,base/sizing,themes/dark,core/shadows base/options,base/sizing --resolveReferences=false",
+    "build-typography-test": "node cli.js tokens.json temp/typography.json base/options,base/sizing,core/typography base/options,base/sizing",
+    "build-composition-test": "node cli.js tokens.json temp/composition.json base/options,base/sizing,core/composition base/options,base/sizing --expandComposition=true",
+    "build-light-test": "node cli.js tokens.json temp/dark.json base/options,base/sizing,themes/dark,core/shadows base/options,base/sizing",
+    "build-dark-test": "node cli.js tokens.json temp/light.json base/options,base/sizing,themes/light,core/shadows base/options,base/sizing",
+    "build-single-file-test-theme-config": "node cli.js tokens.json --theme --themeOutputPath temp",
+    "build-single-file-test-order-config": "node cli.js order.json --theme --themeOutputPath temp",
+    "build-folder-test": "node cli.js tokens temp/folder-light.json base/options,base/sizing,themes/light,core/shadows base/options,base/sizing",
+    "build-folder-test-theme-config": "node cli.js tokens --theme --themeOutputPath temp",
+    "build-resolve-false-test": "node cli.js tokens.json temp/resolve-false.json base/options,base/sizing,themes/dark,core/shadows base/options,base/sizing --resolveReferences=false",
     "test": "npm run build-typography-test && npm run build-light-test && npm run build-dark-test && npm run build-single-file-test-theme-config && npm run build-folder-test && npm run build-folder-test-theme-config && npm run build-resolve-false-test && npm run build-composition-test && npm run build-single-file-test-order-config && jest"
   },
   "author": "six7",
   "license": "MIT",
-  "bin": "index.js",
+  "bin": "cli.js",
   "dependencies": {
     "yargs": "^17.4.1"
   },
   "files": [
     "index.js",
+    "cli.js",
     "dist"
   ],
   "devDependencies": {
